chore: reformat and cleanup code

+++ src/Bancho.Application/PacketHandlers/Spectating/StopSpectatingHandler.cs
@@ -1,7 +1,6 @@
+using Bancho.Application.PacketHandlers.Core;
 using Bancho.Application.Sessions;
 using Bancho.Application.UseCases.Spectating;
-using Bancho.Protocol;
-using Bancho.Application.PacketHandlers.Core;
 using Bancho.Protocol.Packets;
 
 namespace Bancho.Application.PacketHandlers.Spectating;
@@ -16,10 +15,7 @@ public sealed class StopSpectatingHandler(SpectatorService spectatorService) : I
     public Task HandleAsync(PlayerSession player, BanchoPacketReader reader)
     {
         var host = player.Spectating;
-        if (host is not null)
-        {
-            spectatorService.RemoveSpectator(host, player);
-        }
+        if (host is not null) spectatorService.RemoveSpectator(host, player);
 
         return Task.CompletedTask;
     }
